requirements.txt
# Python dependencies
fastapi
uvicorn
pydantic
llama-cpp-python
TTS
sentence-transformers
faiss-cpu
whisper
pytesseract
Pillow
pdf2image
PyMuPDF
langchain
langgraph
